Split TaskHandler monster file
